refactor: transferring items to the components folder
