+++ social-network/src/components/NewsCard/index.ts
@@ -1,0 +1,1 @@
+export {default} from './NewsCard'
