+++ social-network/src/components/NewsCreator/NewsCreator.module.scss
@@ -1,0 +1,42 @@
+.create{
+    display: flex;
+    flex-direction: column;
+    gap: 20px;
+    background-color: #fff;
+    -webkit-transition: box-shadow 300ms cubic-bezier(0.4, 0, 0.2, 1) 0ms;
+    transition: box-shadow 300ms cubic-bezier(0.4, 0, 0.2, 1) 0ms;
+    border-radius: 4px;
+    box-shadow: 0px 2px 1px -1px rgba(0, 0, 0, 0.2), 0px 1px 1px 0px rgba(0, 0, 0, 0.14), 0px 1px 3px 0px rgba(0, 0, 0, 0.12);
+}
+.createHeader {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 5px 10px;
+}
+.createFooter {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 5px 10px;
+}
+.createIcons {
+    display: flex;
+    gap: 25px;
+}
+.createItem {
+    display: flex;
+    gap: 5px;
+    align-items: center;
+    color: var(--gray-color);
+    font-size: 17px;
+    font-weight: 700;  
+    cursor: pointer;
+}
+.contentInput {
+    & > :not(style){
+        width: 65ch;
+        background-color: var(--light-gray);
+        margin: 1;
+    }
+}
