+++ social-network/src/components/NewsCreator/index.ts
@@ -1,0 +1,1 @@
+export { default } from './NewsCreator'
